+++ bfabric/src/bfabric/oauth/__init__.py
@@ -3,7 +3,8 @@
 # Importing this module pulls authlib, joserfc and httpx; ``import bfabric`` alone does not.
 from bfabric.oauth._credential_provider import OAuthCredentialProvider
 from bfabric.oauth._device_code import device_code_login
-from bfabric.oauth._pkce import pkce_login
+from bfabric.oauth._endpoints import authorize_url, token_url
+from bfabric.oauth._pkce import exchange_code, pkce_login
 from bfabric.oauth._registration import register_client, register_webapp
 from bfabric.oauth._token_cache import TokenCache, compute_token_cache_path
 from bfabric.oauth._url_token import UrlTokenContext
@@ -14,9 +15,12 @@ __all__ = [
     "TokenCache",
     "UrlTokenContext",
     "WebappClient",
+    "authorize_url",
     "compute_token_cache_path",
     "device_code_login",
+    "exchange_code",
     "pkce_login",
     "register_client",
     "register_webapp",
+    "token_url",
 ]